Recoded all of peaks and data
Also started pylint cleaning
