+++ maroonx/DRAGONS-X.code-workspace
@@ -1,0 +1,7 @@
+{
+    "folders": [
+        {
+            "path": "../maroonxdr/maroonx"
+        }
+    ]
+}
