--- maroonxdr/doc/usermanuals/MAROONXDR_UserManual/README
@@ -1,1 +1,0 @@
-Location of the Sphinx documents.
