+++ maroonxdr/maroonx/lookups/refwavelengthdb.py
@@ -1,0 +1,8 @@
+# maroonx/refwavelengthdb.py
+#
+# This file contains the reference wavelength lookup table for MaroonX
+
+refwavelength_dict = {
+    "REFWAVELENGTH_b.fits" : "REFWAVELENGTH_b.fits",
+    "REFWAVELENGTH_r.fits" : "REFWAVELENGTH_r.fits"
+}
