+++ maroonxdr/maroonx/maroonx_fit/__init__.py
@@ -1,0 +1,3 @@
+"""
+A basic init file so python knows to treat maroonx_fit as a sub-package
+"""
