+++ maroonxdr/maroonx/maroonx_utils.py
@@ -1,7 +1,11 @@
-import numpy as np
+'''
+MAROONX Utils file.  Contains functions used to load data from reference files.
+'''
+import os
+import logging
 import astrodata
-import pandas as pd
-from . import maroonx_etalon_fit
+import numpy as np
+from .lookups import siddb, maskdb, refwavelengthdb
 
 def load_recordings(ad, guess_file, fibers, orders):
     """
@@ -71,7 +75,7 @@ def load_recordings(ad, guess_file, fibers, orders):
                     # Get each individual 4036 pixel row
                     for fiber_row, flat_row, guess_row \
                     in zip(reduced_fiber, reduced_flat, guess_fiber):
-                        data = fiber_row / flat_row #Normalize according to flat 
+                        data = fiber_row / flat_row #Normalize according to flat
                         # TODO: Check with Andreas if we should compute error too
                         guess_data = guess_row / flat_row # Normalize according to flat
                         guess_data = guess_data / np.nanmedian(guess_data[500:3500])*np.nanmedian(data[500:3500])
@@ -120,158 +124,63 @@ def load_recordings(ad, guess_file, fibers, orders):
                 data = reduced_fiber[i] / reduced_flat[i] #Normalize according to flat
                 yield fiber_number, order, data, None
 
-def insert_peak_parameters(results):
+def get_sid_filename(ad):
     """
-    Extract the parameters for each peak.
+    Gets stripe ID file for input frame.  SID will not be caldb compliant as it is
+    instrument specific.
 
-    Parameters
-    ----------
-    results: FitResults NamedTuple
-        Results of the fits generated by iterative fit
-    
+    Returns
+     -------
+    str/None: Filename of appropriate sid
+    """
+    logging.basicConfig(level=logging.DEBUG)
+    log = logging.getLogger(__name__)
+    arm = ('b' if 'BLUE' in ad.tags else 'r') #Get appropriate arm
+    sid_dir = os.path.join(os.path.dirname(siddb.__file__), 'SID')
+    db_matches = sorted((k, v) for k, v in siddb.sid_dict.items()
+                            if arm in k) #Check if there is a Stripe ID file for the given arm
+    if db_matches:
+        sid = db_matches[-1][1]
+    else:
+        log.warning(f'No SID found for {ad.filename}')
+        return None
+    return sid if sid.startswith(os.path.sep) else \
+        os.path.join(sid_dir, sid)
+
+def get_bpm_filename(ad):
+    """
+    Gets bad pixel mask for input MX science frame.
+    this function can be removed when MX is bpm caldb compliant
+
     Returns
     -------
-    peaks: A pandas dataframe that can be converted to a fits table
+    str/None: Filename of the appropriate bpms
     """
-    peaks = pd.DataFrame()
-    
-    # Making the empty columns
-    fiber_list = []
-    order_list = []
-    offset_list = []
-    center_list = []
-    amplitude_list = []
-    sigma1_list = []
-    sigma2_list = []
-    width_list = []
-    recorded_list = []
-    lq_cost_list = []
-    lq_status_list = []
+    arm = ('b' if 'BLUE' in ad.tags else 'r') #Get appropriate arm
+    bpm_dir = os.path.join(os.path.dirname(maskdb.__file__), 'BPM')
+    bpm = 'BPM_'+arm+'_0000.fits' #Append appropriate arm to the bpm name
+    return bpm if bpm.startswith(os.path.sep) else \
+        os.path.join(bpm_dir, bpm)
 
-    for result in results:
-        offset = maroonx_etalon_fit.get_offset(result.parameters, result.meta_parameters)
-        fits = maroonx_etalon_fit.eval_polynomials_at_centers(result.parameters, result.meta_parameters)
-
-        for values, f in zip(fits.T, result.peak_fit_results):
-            center, amplitude, sigma1, sigma2, width = values
-            
-            # Append the values to the columns
-            fiber_list.append(result.fiber)
-            order_list.append(result.order)
-            offset_list.append(offset)
-            center_list.append(center)
-            amplitude_list.append(amplitude)
-            sigma1_list.append(sigma1)
-            sigma2_list.append(sigma2)
-            width_list.append(width)
-            recorded_list.append(result.recording_time)
-            lq_cost_list.append(f.cost)
-            lq_status_list.append(f.status)
-
-    # Insert the columns into the dataframe
-    peaks['fiber'] = fiber_list
-    peaks['order'] = order_list
-    peaks['offset'] = offset_list
-    peaks['center'] = center_list
-    peaks['amplitude'] = amplitude_list
-    peaks['sigma1'] = sigma1_list
-    peaks['sigma2'] = sigma2_list
-    peaks['width'] = width_list
-    peaks['recorded'] = recorded_list
-    peaks['lq_cost'] = lq_cost_list
-    peaks['lq_status'] = lq_status_list
-
-    return peaks
-
-def insert_polynomial_parameters(results):
+def get_refwavelength_filename(ad):
     """
-    Extract polynomial parameters from results.
+    Gets reference wavelength file for input MX science frame.
+    REF wavelength files are not caldb compliant as they are instrument specific.
 
-    Parameters
-    ----------
-    results: FitResults NamedTuple
-        Results of the fits generated by iterative fit
-    
+    Returns
+    -------
+    str/None: Filename of the appropriate ref wavelength file
     """
-    meta_parameters = results[0].meta_parameters
-    
-    #Check valid meta_parameters.  Do not check total and number of peaks.
-    if not all(result.meta_parameters.offset == meta_parameters.offset for result in results)\
-            and all(result.meta_parameters.sigma == meta_parameters.sigma for result in results)\
-            and all(result.meta_parameters.width == meta_parameters.width for result in results)\
-            and all(result.meta_parameters.use_sigma_lr == meta_parameters.use_sigma_lr for result in results)\
-            and all(result.meta_parameters.use_sigma_lr == meta_parameters.use_sigma_lr for result in results):
-        raise ValueError('All meta_parameters must be the same')
-    
-    if meta_parameters.use_sigma_lr == True:
-        poly = pd.DataFrame()
+    logging.basicConfig(level=logging.DEBUG)
+    log = logging.getLogger(__name__)
+    arm = ('b' if 'BLUE' in ad.tags else 'r') #Get appropriate arm
+    refwavelength_dir = os.path.join(os.path.dirname(refwavelengthdb.__file__), 'REFWAVELENGTH')
+    db_matches = sorted((k, v) for k, v in refwavelengthdb.refwavelength_dict.items()\
+    if arm in k) #Check if there is a reference wavelength file for the given arm
+    if db_matches:
+        refwavelength = db_matches[-1][1]
     else:
-        poly = pd.DataFrame()
-    # Make the empty columns
-    recorded_list = []
-    fiber_list = []
-    order_list = []
-    fitrange_list = []
-    offset_list = []
-    lq_cost_list = []
-    lq_status_list = []
-    if meta_parameters.use_sigma_lr == True:
-        sigma_l_coefficients_list = []
-        sigma_r_coefficients_list = []
-    else:
-        sigma_coefficients_list = []
-    width_coefficients_list = []
-
-    # Get the polynomial parameters and insert into dataframe
-    for result in results:
-        values = maroonx_etalon_fit.split_parameters(result.parameters, result.meta_parameters)
-        offset, p_sigma_l, p_sigma_r, p_width, _, _ = values
-
-        f = result.polynomial_fit_result
-
-        # Append the values to the columns
-        recorded_list.append(result.recording_time)
-        fiber_list.append(result.fiber)
-        order_list.append(result.order)
-
-        # Convert the fitrange to a string to store in the fits table
-        fitrange = [result.x[0], result.x[-1] + 1]
-        fitrange_str = ','.join([str(x) for x in fitrange])
-        fitrange_list.append(fitrange_str)
-
-        offset_list.append(offset)
-        lq_cost_list.append(f.cost)
-        lq_status_list.append(f.status)
-        if meta_parameters.use_sigma_lr == True:
-            # Convert the list of coefficients to a string to store in the fits table
-            sigma_l_coefficients_str = ','.join([str(x) for x in list(p_sigma_l)])
-            sigma_r_coefficients_str = ','.join([str(x) for x in list(p_sigma_r)])
-            sigma_l_coefficients_list.append(sigma_l_coefficients_str)
-            sigma_r_coefficients_list.append(sigma_r_coefficients_str)
-
-        else:
-            sigma_coefficients_str = ','.join([str(x) for x in list(p_sigma_l)])
-            sigma_coefficients_list.append(sigma_coefficients_str)
-    
-        width_coefficients_str = ','.join([str(x) for x in list(p_width)])
-        width_coefficients_list.append(width_coefficients_str)
-
-    # Insert the columns into the dataframe
-    poly['recorded'] = recorded_list
-    poly['fiber'] = fiber_list
-    poly['order'] = order_list
-    poly['fitrange'] = fitrange_list
-    poly['offset'] = offset_list
-    poly['lq_cost'] = lq_cost_list
-    poly['lq_status'] = lq_status_list
-    if meta_parameters.use_sigma_lr == True:
-        poly['sigma_l_coefficients'] = sigma_l_coefficients_list
-        poly['sigma_r_coefficients'] = sigma_r_coefficients_list
-    else:
-        poly['sigma_coefficients'] = sigma_coefficients_list
-    poly['width_coefficients'] = width_coefficients_list
-
-    return poly
-
-
-        
+        log.warning(f'No reference wavelength file found for {ad.filename}')
+        return None
+    return refwavelength if refwavelength.startswith(os.path.sep) else \
+        os.path.join(refwavelength_dir, refwavelength)
