+++ maroonxdr/maroonx/tests/complete/dark_test.py
@@ -9,23 +9,26 @@ and to have got the test fits files from Kathleen.
 """
 import glob
 import sys
-import astrodata
+from pathlib import Path
 from recipe_system.reduction.coreReduce import Reduce
 from gempy.adlibrary import dataselect
+import astrodata
 
-from pathlib import Path
 parent_dir = Path(__file__).parents[4]
 sys.path.append(str(parent_dir))
 import maroonx_instruments
 
 # Get all files in the science_dir.  Change the path here to suit your installation.
-all_files = glob.glob('/Users/rohan/Desktop/DRAGONS-X/MAROONXDR/science_dir/*_r_0300.fits')
+all_files = glob.glob('/Users/rohan/Desktop/DRAGONS-X/MAROONXDR/science_dir/*_DDDDE_r_*.fits')
 
 all_files.sort()
-just_darks = dataselect.select_data(all_files, ['DARK'])
 
+just_darks = dataselect.select_data(all_files, tags=['DARK'])
+print(just_darks)
 # Run reduce on all selected files
 myreduce = Reduce()
-myreduce.files.extend(just_darks)
+print(myreduce.files)
+
+myreduce.files.extend(all_files)
 myreduce.drpkg= 'maroonxdr'
 myreduce.runr()
