+++ maroonxdr/maroonx/tests/image/test_file_sorting.py
@@ -9,7 +9,7 @@ import numpy as np
 
 parent_dir = Path(__file__).parents[4]
 sys.path.append(str(parent_dir))
-from maroonxdr.maroonx.primitives_maroonx import MAROONX
+from MAROONXDR.maroonxdr.maroonx.primitives_maroonx.primitives_maroonx_generic import MAROONX
 import maroonx_instruments
 
 os.chdir(parent_dir)
