+++ maroonxdr/maroonx/tests/image/test_stray_light_removal.py
@@ -8,7 +8,7 @@ import astrodata
 
 parent_dir = Path(__file__).parents[4]
 sys.path.append(str(parent_dir))
-from maroonxdr.maroonx.primitives_maroonx import MAROONX
+from MAROONXDR.maroonxdr.maroonx.primitives_maroonx.primitives_maroonx_generic import MAROONX
 import maroonx_instruments
 
 #Change the path names below to suit your installation and the files given to you
